VideoStream moved in separate file; config manager implemented;
