+++ .env.example
@@ -1,0 +1,9 @@
+# Email you want to send the update from (only works with gmail)
+FROM_EMAIL='xxx@gmail.com'
+
+# You can generate an app password here to avoid storing your password in plain text
+# https://support.google.com/accounts/answer/185833?hl=en
+FROM_EMAIL_PASSWORD='xxxx'
+
+# Email you want to send the update to
+TO_EMAIL='xxx@gmail.com'
