+++ Pipfile
@@ -7,6 +7,8 @@ name = "pypi"
 opencv-python = "*"
 numpy = "*"
 tflite-runtime = "*"
+loguru = "*"
+python-decouple = "*"
 
 [dev-packages]
